Push pipeline changes
Pipeline read message block pointer addition.  Zero buffer reads in the read/write handler to detect disconnected socket when the handleReadData doesn't lead to a call to drain the channel to a buffer.  This probably doesn't work though, because the zero buffer does not even attempt to read data from the channel.  Is there another way of updating the state of the non-blocking channel so that a channel shut can be detected.